Changed WAIT behavior & Fixed no sound issue
Changed the way WAIT time is handled in the Scanner App.

Previously, the WAIT timer would start counting down the moment a signal was locked onto. This would cause the scanner to start searching for signals again at the end of the timer even if there is still transmission occurring on the current channel.

This change makes it so the WAIT timer only counts down after the signal has dropped back below the Squelch level. If the transmission starts back up again before the WAIT timer reaches zero, it resets the timer and will start again when the signal drops back below the Squelch level again. Only when the signal is below the Squelch level for the WAIT time will it start scanning again.

--------------------------------------

Also fixed the issue of no sound output that was effecting Portapack H2 devices with the WM8731 audio chip.